tests: Move v1 route tests from test_service.py to routes_v1.py
test_service.py now implements its own setUpModule and tearDownModule
funxtions and imports routes_v1 in order to create an instantiable
TestCase subclass in the same module (without having to have all
the tests inline)
